requirements.txt
botpy
pyyaml
